# This is a combination of 10 commits. # This is the 1st commit message:
ACME client & server iteration

# This is the commit message #2:

fix http challenge urls. clean up comments

# This is the commit message #3:

Add webapp to public_html. more misc cleanup

# This is the commit message #4:

support configurable / dynamic port assignments for http based challenges

# This is the commit message #5:

misc bug fixes and cleanup

# This is the commit message #6:

- Add key rotation support to key stores
- Add overwrite flag to key store backend and save operations
- Add docker and packer support for the core platform and RPI
- Create Trusted Platform ISO build scripts
- Re-organize webservice package, decouple configs
- Add ACME client cross-signing support
- Create examples directory

# This is the commit message #7:

Remove TSS attestation workflow from core project, move to examples

# This is the commit message #8:

Remove protobuf from core project, relocate to TSS examples

# This is the commit message #9:

- Add support for external ACME servers (tested w/ LetsEncrypt)
- Add DNS server w/ support for internal and public DNS zones
- Add full support for dns-01 ACME challenge w/ DNS server integration
- Create docker based ISO builders

# This is the commit message #10:

- Add support for automated ACME certificates deployed to web server
  via external CA's (e.g. LetsEncrypt)
- Add virtual hosts, url rewrite and reverse proxy support to web server
- Add support for HTTP/3 QUIC
- Create build-public-html target to automate dashboard build
- Re-organize configs, cleanup
- Create custom security logging facility
- Add ACME docs, lots of cleanup
- Fix bugs and broken tests
- Clean up VS code tasks and targets
- Iterate on minimal docker and ISO build targets for platform and ansible

+++ pkg/store/keystore/backend_file.go
@@ -25,7 +25,7 @@ var (
 
 type KeyBackend interface {
 	Get(attrs *KeyAttributes, extension FSExtension) ([]byte, error)
-	Save(attrs *KeyAttributes, data []byte, extension FSExtension) error
+	Save(attrs *KeyAttributes, data []byte, extension FSExtension, overwrite bool) error
 	Delete(attrs *KeyAttributes) error
 }
 
@@ -94,12 +94,20 @@ func NewFileBackend(
 func (fb *FileBackend) Save(
 	attrs *KeyAttributes,
 	data []byte,
-	extension FSExtension) error {
+	extension FSExtension,
+	overwrite bool) error {
 
 	file, err := fb.fileNameFromKeyAttributes(attrs, extension)
 	if err != nil {
 		fb.logger.Errorf("%s: %s", err, file)
 		return err
+	}
+	if overwrite {
+		if err = afero.WriteFile(fb.fs, file, data, 0644); err != nil {
+			fb.logger.Errorf("%s: %s", err, file)
+			return err
+		}
+		return nil
 	}
 	if _, err := fb.fs.Stat(file); errors.Is(err, os.ErrNotExist) {
 		if err = afero.WriteFile(fb.fs, file, data, 0644); err != nil {
